requirement.txt
requests
Pillow
PrettyTable
colorama


namedtupled
PyYAML

redis # use redis for storing cookies
ipython # for debug
